Refactor measurements modal: replace MeasurementsModalContent with Measurements component, enhancing structure and functionality. Update imports in mobile and product info components accordingly.
