+++ src/app/product/[...productParams]/_components/product-info.tsx
@@ -7,7 +7,7 @@ import { cn } from "@/lib/utils";
 import Modal from "@/components/ui/modal";
 import { Text } from "@/components/ui/text";
 
-import { MeasurementsModalContent } from "./measurements-modal-content";
+import { Measurements } from "./measurements";
 import { AddToCartForm } from "./select-size-add-to-cart/index";
 import { useData } from "./select-size-add-to-cart/useData";
 
@@ -37,8 +37,8 @@ export function ProductInfo({ product }: { product: common_ProductFull }) {
 
   return (
     <div className="border-inactive absolute bottom-2.5 right-2.5 h-[245px] w-[600px] border bg-bgColor p-2.5">
-      <div className="relative grid h-full grid-cols-2 gap-x-5">
-        <div className={"flex flex-col justify-between"}>
+      <div className="grid h-full grid-cols-2 gap-x-5">
+        <div className="flex flex-col justify-between">
           <Text
             variant="uppercase"
             className={cn("", {
@@ -92,7 +92,7 @@ export function ProductInfo({ product }: { product: common_ProductFull }) {
               </Modal>
             </div>
             <Modal openElement="size guide">
-              <MeasurementsModalContent
+              <Measurements
                 id={productId}
                 sizes={sizes}
                 categoryId={categoryId}
